requirements.txt
pyTorch==2.3.0
sklearn==1.5.1
nltk==3.8.1
numpy==1.24.0

